Filtering works. Working example is the review component. Wasted a lot´s of time for the review carousel, instead for the filtering but hey.
IMPORTANT: Classes are now in api.ts and dashboard.ts

Wasted a lot of time for the review carousel to work, because
materialize sucks at handling dynamic changes. For now we have a dirty
workaround (yeah!)
+ Selection now easy and implemented. Next step is to implement update
functionality for every component and to update the data service. I can
do this later...